+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Qristal is a full-stack SDK for quantum accelerators.
 ### Added 
 
 - Added functions to convert quantum process matrices in their Choi representation to and from their superoperator representation. Also added convenient function delegations involving standard process matrices (chi), and Kraus representations. 
+- Added a custom pyGSTi model pack to the pyGSTi python examples for gate set tomography protocols using the QDK's native gate set: Rx(pi/2), Ry(pi/2), and CZ. 
+- Adjusted the pyGSTi_runner.cpp and execute_GST_XYCZ.py examples to use the new custom model pack.
 
 ## [1.6.0] - 2024-06-18